rpa-system/automation/production_automation_service.py
#!/usr/bin/env python3
"""
Production-Ready Automation Service with Enhanced Kafka Support
This version handles both Kafka and HTTP-only modes gracefully
"""

import os
import sys
import time
import logging
import threading
import signal
import json
import uuid
from flask import Flask, request, jsonify

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Flask app
app = Flask(__name__)

# Global state
kafka_available = False
shutdown_event = threading.Event()

class SimpleKafkaManager:
    """
    Simplified Kafka manager that gracefully handles missing dependencies
    """
    
    def __init__(self):
        self.kafka_enabled = os.getenv('KAFKA_ENABLED', 'true').lower() == 'true'
        self.bootstrap_servers = os.getenv('KAFKA_BOOTSTRAP_SERVERS', 'localhost:9092')
        self.task_topic = os.getenv('KAFKA_TASK_TOPIC', 'automation-tasks')
        self.result_topic = os.getenv('KAFKA_RESULT_TOPIC', 'automation-results')
        self.consumer_group = os.getenv('KAFKA_CONSUMER_GROUP', 'automation-workers')
        
        self.producer = None
        self.consumer = None
        self.is_connected = False
        
        logger.info(f"🔧 Kafka Manager - Enabled: {self.kafka_enabled}")
    
    def initialize(self):
        """Initialize Kafka connections if available"""
        if not self.kafka_enabled:
            logger.info("🔇 Kafka disabled - running in HTTP-only mode")
            return True
        
        try:
            # Try to import kafka
            from kafka import KafkaProducer, KafkaConsumer
            from kafka.errors import NoBrokersAvailable
            
            # Test basic connectivity first
            import socket
            host, port = self.bootstrap_servers.split(':')
            sock = socket.create_connection((host, int(port)), timeout=5)
            sock.close()
            
            # Initialize producer
            self.producer = KafkaProducer(
                bootstrap_servers=self.bootstrap_servers.split(','),
                value_serializer=lambda x: json.dumps(x).encode('utf-8'),
                key_serializer=lambda x: x.encode('utf-8') if x else None,
                retries=3,
                request_timeout_ms=10000
            )
            
            # Initialize consumer
            self.consumer = KafkaConsumer(
                self.task_topic,
                bootstrap_servers=self.bootstrap_servers.split(','),
                group_id=self.consumer_group,
                value_deserializer=lambda x: json.loads(x.decode('utf-8')),
                auto_offset_reset='latest',
                enable_auto_commit=True,
                consumer_timeout_ms=1000
            )
            
            self.is_connected = True
            logger.info("✅ Kafka connections established successfully")
            return True
            
        except ImportError:
            logger.warning("⚠️ kafka-python not available - install with: pip install kafka-python==2.0.2")
            return False
        except Exception as e:
            logger.warning(f"⚠️ Kafka connection failed: {e}")
            logger.info("🔧 Service will run in HTTP-only mode")
            return False
    
    def send_result(self, task_id, result, status='completed'):
        """Send result to Kafka if available"""
        if not self.is_connected or not self.producer:
            logger.debug(f"📤 Result for task {task_id} - Kafka not available")
            return False
        
        try:
            message = {
                'task_id': task_id,
                'status': status,
                'result': result,
                'timestamp': time.time(),
                'worker_id': os.getenv('HOSTNAME', 'automation-worker')
            }
            
            future = self.producer.send(self.result_topic, key=task_id, value=message)
            future.get(timeout=10)
            logger.info(f"✅ Result sent to Kafka for task {task_id}")
            return True
            
        except Exception as e:
            logger.error(f"❌ Failed to send result to Kafka: {e}")
            return False
    
    def start_consumer_loop(self):
        """Start Kafka consumer loop in background"""
        if not self.is_connected or not self.consumer:
            logger.info("🔇 Kafka consumer not started - not available")
            return
        
        def consumer_loop():
            logger.info(f"👂 Starting Kafka consumer for topic: {self.task_topic}")
            
            try:
                while not shutdown_event.is_set():
                    try:
                        message_batch = self.consumer.poll(timeout_ms=1000)
                        
                        for topic_partition, messages in message_batch.items():
                            for message in messages:
                                task_data = message.value
                                task_id = task_data.get('task_id', str(uuid.uuid4()))
                                
                                logger.info(f"📨 Processing Kafka task: {task_id}")
                                
                                # Process task (implement your task processing here)
                                result = self.process_task(task_data)
                                
                                # Send result back
                                self.send_result(task_id, result)
                                
                    except Exception as e:
                        logger.error(f"❌ Error in consumer loop: {e}")
                        time.sleep(5)
                        
            except Exception as e:
                logger.error(f"❌ Consumer loop failed: {e}")
            
            logger.info("🛑 Kafka consumer loop stopped")
        
        # Start consumer thread
        consumer_thread = threading.Thread(target=consumer_loop, daemon=True)
        consumer_thread.start()
        logger.info("🎧 Kafka consumer thread started")
    
    def process_task(self, task_data):
        """Process automation task"""
        task_type = task_data.get('task_type', 'unknown')
        
        try:
            if task_type == 'web_automation':
                return {'success': True, 'message': f'Processed {task_type}', 'data': task_data}
            elif task_type == 'form_submission':
                return {'success': True, 'message': f'Processed {task_type}', 'data': task_data}
            elif task_type == 'data_extraction':
                return {'success': True, 'message': f'Processed {task_type}', 'data': task_data}
            else:
                return {'success': False, 'error': f'Unknown task type: {task_type}'}
                
        except Exception as e:
            logger.error(f"❌ Task processing failed: {e}")
            return {'success': False, 'error': str(e)}
    
    def shutdown(self):
        """Graceful shutdown"""
        logger.info("🛑 Shutting down Kafka connections...")
        
        if self.producer:
            try:
                self.producer.flush()
                self.producer.close()
                logger.info("✅ Kafka producer closed")
            except Exception as e:
                logger.error(f"❌ Error closing producer: {e}")
        
        if self.consumer:
            try:
                self.consumer.close()
                logger.info("✅ Kafka consumer closed")
            except Exception as e:
                logger.error(f"❌ Error closing consumer: {e}")
    
    def get_status(self):
        """Get Kafka status"""
        return {
            'enabled': self.kafka_enabled,
            'connected': self.is_connected,
            'bootstrap_servers': self.bootstrap_servers,
            'task_topic': self.task_topic,
            'result_topic': self.result_topic
        }

# Initialize Kafka manager
kafka_manager = SimpleKafkaManager()

def signal_handler(signum, frame):
    """Handle shutdown signals"""
    logger.info(f"📡 Received signal {signum}, shutting down...")
    shutdown_event.set()
    kafka_manager.shutdown()
    sys.exit(0)

# Register signal handlers
signal.signal(signal.SIGTERM, signal_handler)
signal.signal(signal.SIGINT, signal_handler)

# Flask Routes
@app.route('/health', methods=['GET'])
def health_check():
    """Enhanced health check with Kafka status"""
    return jsonify({
        'status': 'healthy',
        'service': 'automation',
        'timestamp': time.time(),
        'uptime': time.time() - getattr(app, 'start_time', time.time()),
        'kafka': kafka_manager.get_status(),
        'capabilities': ['web_automation', 'form_submission', 'data_extraction', 'file_download']
    }), 200

@app.route('/automation/execute', methods=['POST'])
def execute_automation():
    """Direct automation execution endpoint"""
    try:
        task_data = request.get_json()
        
        if not task_data:
            return jsonify({'error': 'No task data provided'}), 400
        
        task_type = task_data.get('task_type')
        if not task_type:
            return jsonify({'error': 'No task_type specified'}), 400
        
        # Add task ID
        task_id = task_data.get('task_id', str(uuid.uuid4()))
        task_data['task_id'] = task_id
        
        # Process task directly
        result = kafka_manager.process_task(task_data)
        
        return jsonify({
            'success': True,
            'task_id': task_id,
            'task_type': task_type,
            'result': result,
            'timestamp': time.time(),
            'processing_mode': 'direct_http'
        }), 200
        
    except Exception as e:
        logger.error(f"❌ Direct automation execution failed: {e}")
        return jsonify({
            'success': False,
            'error': str(e),
            'timestamp': time.time()
        }), 500

@app.route('/automation/status', methods=['GET'])
def get_status():
    """Get service status"""
    return jsonify({
        'service': 'automation',
        'version': '2.0.0',
        'status': 'running',
        'kafka': kafka_manager.get_status(),
        'environment': {
            'kafka_enabled': os.getenv('KAFKA_ENABLED', 'true'),
            'bootstrap_servers': os.getenv('KAFKA_BOOTSTRAP_SERVERS', 'localhost:9092'),
            'port': os.getenv('PORT', '7001')
        }
    }), 200

if __name__ == '__main__':
    # Set startup time
    app.start_time = time.time()
    
    logger.info("🚀 Starting EasyFlow Automation Service...")
    
    # Initialize Kafka
    kafka_available = kafka_manager.initialize()
    
    if kafka_available:
        # Start Kafka consumer
        kafka_manager.start_consumer_loop()
        logger.info("✅ Service started with Kafka support")
    else:
        logger.info("⚠️ Service started in HTTP-only mode")
    
    # Start Flask app
    port = int(os.getenv('PORT', 7001))
    host = os.getenv('HOST', '0.0.0.0')
    
    logger.info(f"🌐 Starting server on {host}:{port}")
    
    try:
        app.run(host=host, port=port, debug=False, threaded=True)
    except KeyboardInterrupt:
        logger.info("⌨️ Shutting down...")
    finally:
        kafka_manager.shutdown()
