rpa-system/automation/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
selenium==4.23.1
webdriver-manager==4.0.2
requests==2.32.3
python-magic==0.4.27
cryptography==42.0.8
kafka-python==2.0.2
asyncio-throttle==1.0.2
